+++ components/selector/src/Option.js
@@ -5,11 +5,12 @@ import PropTypes from 'prop-types';
 
 import { defaultOptionBtn } from './overrides/optionBtn';
 import { defaultOption } from './overrides/option';
-import { defaultContent } from './overrides/content';
 import { defaultPictogram } from './overrides/pictogram';
 import { defaultIcon } from './overrides/icon';
 import { defaultText } from './overrides/text';
 import { defaultLabel } from './overrides/label';
+import { defaultLabelPrimary } from './overrides/labelPrimary';
+import { defaultLabelSecondary } from './overrides/labelSecondary';
 import { defaultHint } from './overrides/hint';
 import { defaultIndicator } from './overrides/indicator';
 
@@ -24,6 +25,7 @@ export const Option = ({
 	value,
 	pictogram,
 	icon,
+	secondaryLabel,
 	checked: checkedProp,
 	hint,
 	className,
@@ -60,11 +62,12 @@ export const Option = ({
 	const defaultOverrides = {
 		Option: defaultOption,
 		OptionBtn: defaultOptionBtn,
-		Content: defaultContent,
 		Pictogram: defaultPictogram,
 		Icon: defaultIcon,
 		Text: defaultText,
 		Label: defaultLabel,
+		LabelPrimary: defaultLabelPrimary,
+		LabelSecondary: defaultLabelSecondary,
 		Hint: defaultHint,
 		Indicator: defaultIndicator,
 	};
@@ -77,6 +80,7 @@ export const Option = ({
 		value,
 		pictogram,
 		icon,
+		secondaryLabel,
 		nextIndicator,
 		iconSize,
 		pictogramWidth,
@@ -96,11 +100,20 @@ export const Option = ({
 	const {
 		Option: { component: Option, styles: optionStyles, attributes: optionAttributes },
 		OptionBtn: { component: OptionBtn, styles: optionBtnStyles, attributes: optionBtnAttributes },
-		Content: { component: Content, styles: contentStyles, attributes: contentAttributes },
 		Pictogram: { component: Pictogram, styles: pictogramStyles, attributes: pictogramAttributes },
 		Icon: { component: Icon, styles: iconStyles, attributes: iconAttributes },
 		Text: { component: Text, styles: textStyles, attributes: textAttributes },
 		Label: { component: Label, styles: labelStyles, attributes: labelAttributes },
+		LabelPrimary: {
+			component: LabelPrimary,
+			styles: labelPrimaryStyles,
+			attributes: labelPrimaryAttributes,
+		},
+		LabelSecondary: {
+			component: LabelSecondary,
+			styles: labelSecondaryStyles,
+			attributes: labelSecondaryAttributes,
+		},
 		Hint: { component: Hint, styles: hintStyles, attributes: hintAttributes },
 		Indicator: { component: Indicator, styles: indicatorStyles, attributes: indicatorAttributes },
 	} = overrideReconciler(defaultOverrides, tokenOverrides, brandOverrides, componentOverrides);
@@ -162,28 +175,41 @@ export const Option = ({
 				}}
 			/>
 			<OptionBtn state={state} {...optionBtnAttributes(state)} css={optionBtnStyles(state)}>
-				<Content state={state} {...contentAttributes(state)} css={contentStyles(state)}>
-					{pictogram ? (
-						<Pictogram
-							pictogram={pictogram}
+				{pictogram ? (
+					<Pictogram
+						pictogram={pictogram}
+						state={state}
+						{...pictogramAttributes(state)}
+						css={pictogramStyles(state)}
+					/>
+				) : icon ? (
+					<Icon icon={icon} state={state} {...iconAttributes(state)} css={iconStyles(state)} />
+				) : null}
+				<Text state={state} {...textAttributes(state)} css={textStyles(state)}>
+					<Label state={state} {...labelAttributes(state)} css={labelStyles(state)}>
+						<LabelPrimary
 							state={state}
-							{...pictogramAttributes(state)}
-							css={pictogramStyles(state)}
-						/>
-					) : icon ? (
-						<Icon icon={icon} state={state} {...iconAttributes(state)} css={iconStyles(state)} />
-					) : null}
-					<Text state={state} {...textAttributes(state)} css={textStyles(state)}>
-						<Label state={state} {...labelAttributes(state)} css={labelStyles(state)}>
+							{...labelPrimaryAttributes(state)}
+							css={labelPrimaryStyles(state)}
+						>
 							{children}
-						</Label>
-						{hint && (
-							<Hint state={state} {...hintAttributes(state)} css={hintStyles(state)}>
-								{hint}
-							</Hint>
+						</LabelPrimary>
+						{secondaryLabel && (
+							<LabelSecondary
+								state={state}
+								{...labelSecondaryAttributes(state)}
+								css={labelSecondaryStyles(state)}
+							>
+								{secondaryLabel}
+							</LabelSecondary>
 						)}
-					</Text>
-				</Content>
+					</Label>
+					{hint && (
+						<Hint state={state} {...hintAttributes(state)} css={hintStyles(state)}>
+							{hint}
+						</Hint>
+					)}
+				</Text>
 				<Indicator state={state} {...indicatorAttributes(state)} css={indicatorStyles(state)} />
 			</OptionBtn>
 		</Option>
@@ -214,6 +240,11 @@ Option.propTypes = {
 	 * Icon graphic
 	 */
 	icon: PropTypes.func,
+
+	/**
+	 * Secondary label text
+	 */
+	secondaryLabel: PropTypes.node,
 
 	/**
 	 * Check the Selector option
@@ -266,11 +297,6 @@ Option.propTypes = {
 			component: PropTypes.elementType,
 			attributes: PropTypes.func,
 		}),
-		Content: PropTypes.shape({
-			styles: PropTypes.func,
-			component: PropTypes.elementType,
-			attributes: PropTypes.func,
-		}),
 		Pictogram: PropTypes.shape({
 			styles: PropTypes.func,
 			component: PropTypes.elementType,
@@ -287,6 +313,16 @@ Option.propTypes = {
 			attributes: PropTypes.func,
 		}),
 		Label: PropTypes.shape({
+			styles: PropTypes.func,
+			component: PropTypes.elementType,
+			attributes: PropTypes.func,
+		}),
+		LabelPrimary: PropTypes.shape({
+			styles: PropTypes.func,
+			component: PropTypes.elementType,
+			attributes: PropTypes.func,
+		}),
+		LabelSecondary: PropTypes.shape({
 			styles: PropTypes.func,
 			component: PropTypes.elementType,
 			attributes: PropTypes.func,