+++ components/selector/src/overrides/text.js
@@ -14,6 +14,8 @@ const Text = ({ state: _, ...rest }) => <div {...rest} />;
 
 const textStyles = () => ({
 	label: getLabel('selector-option-text'),
+	flex: 1,
+	minWidth: 0, //important for text truncation with ellipsis
 });
 
 // ==============================