+++ components/badge/CHANGELOG.md
@@ -1,6 +1,14 @@
 # @westpac/badge
 
+## 1.0.2
+
+### Patch Changes
+
+- Updated dependencies [6a9ef63b]
+  - @westpac/core@2.0.0
+
 ## 1.0.1
+
 ### Patch Changes
 
 - 520c2fc4: fix distribution files