+++ components/hooks/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @westpac/hooks
 
+## 1.1.1
+
+### Patch Changes
+
+- Updated dependencies [6a9ef63b]
+  - @westpac/core@2.0.0
+
 ## 1.1.0
 
 ### Minor Changes