+++ components/list-group/CHANGELOG.md
@@ -1,6 +1,16 @@
 # @westpac/list-group
 
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [6a9ef63b]
+- Updated dependencies [349a4894]
+  - @westpac/core@2.0.0
+  - @westpac/list@1.0.1
+
 ## 1.0.0-beta.2
+
 ### Patch Changes
 
 - 520c2fc4: fix distribution files