+++ components/selector/CHANGELOG.md
@@ -1,1 +1,22 @@
 # @westpac/selector
+
+## 1.1.0
+
+### Minor Changes
+
+- b668515c: - Added secondary label feature with ellipsis truncation
+  - Top align icon/pictogram and indicator
+  - Reduce indicator gap
+  - Reduce icon and pictogram default size
+  - Reduce next indicator size
+  - Remove icon/pictogram vertical stacking in mobile breakpoint and remove content wrapping
+  - Fix hover styling when hovered on indicator
+  - Rename internal label sub-components
+
+## 1.0.0
+
+### Patch Changes
+
+- Updated dependencies [6a9ef63b]
+  - @westpac/core@2.0.0
+  - @westpac/icon@1.3.1